moved: init components to atlassian-plugin.xml
