+++ src/main/java/onlyoffice/utils/parsing/ParsingUtilImpl.java
@@ -1,12 +1,8 @@
 package onlyoffice.utils.parsing;
 
-import javax.enterprise.inject.Default;
-import javax.inject.Named;
 import java.io.InputStream;
 import java.util.Scanner;
 
-@Named
-@Default
 public class ParsingUtilImpl implements ParsingUtil {
     public String getBody(final InputStream stream) {
         Scanner scanner = null;
