+++ src/main/resources/atlassian-plugin.xml
@@ -132,6 +132,45 @@
         <styleClass>onlyoffice-doceditor</styleClass>
     </web-item>
 
+    <component name="Auth Context" key="onlyoffice-auth-context" class="onlyoffice.managers.auth.AuthContextImpl" public="true">
+        <interface>onlyoffice.managers.auth.AuthContext</interface>
+    </component>
+    <component name="Configuration Manager" key="onlyoffice-configuration-manager" class="onlyoffice.managers.configuration.ConfigurationManagerImpl" public="true">
+        <interface>onlyoffice.managers.configuration.ConfigurationManager</interface>
+    </component>
+    <component name="Convert Manager" key="onlyoffice-convert-manager" class="onlyoffice.managers.convert.ConvertManagerImpl" public="true">
+        <interface>onlyoffice.managers.convert.ConvertManager</interface>
+    </component>
+    <component name="Document Manager" key="onlyoffice-document-manager" class="onlyoffice.managers.document.DocumentManagerImpl" public="true">
+        <interface>onlyoffice.managers.document.DocumentManager</interface>
+    </component>
+    <component name="JWT Manager" key="onlyoffice-jwt-manager" class="onlyoffice.managers.jwt.JwtManagerImpl" public="true">
+        <interface>onlyoffice.managers.jwt.JwtManager</interface>
+    </component>
+    <component name="URL Manager" key="onlyoffice-url-manager" class="onlyoffice.managers.url.UrlManagerImpl" public="true">
+        <interface>onlyoffice.managers.url.UrlManager</interface>
+    </component>
+    <component name="Attachment Util" key="onlyoffice-attachment-util" class="onlyoffice.utils.attachment.AttachmentUtilImpl" public="true">
+        <interface>onlyoffice.utils.attachment.AttachmentUtil</interface>
+    </component>
+    <component name="Parsing Util" key="onlyoffice-parsing-util" class="onlyoffice.utils.parsing.ParsingUtilImpl" public="true">
+        <interface>onlyoffice.utils.parsing.ParsingUtil</interface>
+    </component>
+
+    <component-import key="applicationConfiguration" interface="com.atlassian.config.ApplicationConfiguration"/>
+    <component-import key="pluginSettingsFactory" interface="com.atlassian.sal.api.pluginsettings.PluginSettingsFactory"/>
+    <component-import key="localeManager" interface="com.atlassian.confluence.languages.LocaleManager"/>
+    <component-import key="i18nResolver" interface="com.atlassian.sal.api.message.I18nResolver"/>
+    <component-import key="settingsManager" interface="com.atlassian.confluence.setup.settings.SettingsManager"/>
+    <component-import key="attachmentManager" interface="com.atlassian.confluence.pages.AttachmentManager"/>
+    <component-import key="pageManager" interface="com.atlassian.confluence.pages.PageManager"/>
+    <component-import key="transactionTemplate" interface="com.atlassian.sal.api.transaction.TransactionTemplate"/>
+    <component-import key="bootstrapManager" interface="com.atlassian.confluence.setup.BootstrapManager"/>
+    <component-import key="userManager" interface="com.atlassian.sal.api.user.UserManager"/>
+    <component-import key="webResourceUrlProvider" interface="com.atlassian.plugin.webresource.WebResourceUrlProvider"/>
+    <component-import key="formatSettingsManager" interface="com.atlassian.confluence.core.FormatSettingsManager"/>
+    <component-import key="sysInfoService" interface="com.atlassian.confluence.status.service.SystemInformationService"/>
+
     <servlet key="OnlyOfficeDocEditor" class="onlyoffice.OnlyOfficeEditorServlet" name="Document Editor">
         <description>A full-featured editor for the most known formats of text documents, spreadsheets and presentations that can open these types of documents for editing or preview.</description>
         <url-pattern>/onlyoffice/doceditor</url-pattern>
